feat: Enhance SQL schema management and data processing utilities
- Added new schemas for validating BT footfall data across various boundary types, including highstreets, town centres, BIDs, and bespoke areas, ensuring data quality and consistency.
- Updated SQLManager to include additional search paths for SQL files, improving query organization.
- Enhanced DataWriter to support conditional table truncation and loading, optimizing data insertion into PostgreSQL.
- Improved SFTPClient with optimized file transfer techniques based on file size and type, enhancing performance during data processing.
- Introduced new methods in mcard_transform.py for loading CPI data into PostgreSQL and executing optimized SQL adjustments for Mastercard data.
- Refactored SQL queries for better clarity and performance, including updates to existing queries for international transaction data.

+++ highstreets/sql/queries/mcard/weekly/bid_intl_query.sql
@@ -19,7 +19,7 @@ WITH week_data AS (
             )), 'YYYY-MM-DD') AS week_start
     FROM econ_busyness_mcard_raw_18_zoom AS main
     JOIN econ_busyness_mcard_bids_quad_lookup AS lookup 
-      ON main.quad_id = lookup.quad_id
+      ON main.quad_id::bigint = lookup.quad_id::bigint  
     WHERE main.industry = 'Total Retail'
       AND main.segment = 'International'
       AND main.geo_name = 'London'
